// frontend/src/components/orchestrator/types.ts
export interface PatientInfo {
  name: string;
  abhaId: string;
  dob: string;
  gender: string;
  bloodType: string;
  policyNumber: string;
  planType: string;
  residence: string;
  avatarUrl: string;
}